Add a 'VisibilityType' flag to MRTG. Refs #439.
This is a first step towards having a formal separation between the 'compiler'
identity of a function and its 'type' identity, which can be less agressive and
which needs to be stable during module construction.

As part of this, rename 'identityHash' to 'compilerHash' and add flags
to python functions to control the style of hash we want. At this point,
the two hashes are the same.