More carefully encapsulate the compiler cache within NativeCompiler.
We want for NativeCompiler to provide a simple, clean abstraction for
a code-compilation service. You should be able to push code in, ask whether
you've pushed code before, and get TypedCallTargets and function pointers
out, and not really need to worry about the compiler cache.

This change moves the compiler cache fully behind the native compiler,
and tries to ensure that the python converter doesn't itself need to
track the functions that its compiled in the past. Instead, we're moving
towards it being stateless across invocations (and we'll probably do
the same thing for the native-to-llvm converter as well) since its
substantially simpler to think about.

As part of this change, we modify the compiler cache to acknowledge that
modules inside might not be loadable and that as a result, it might think
that a symbol is available when it is not. In this pathway, the compiler
cache tracks all modules that define a symbol, as well as the first module
that it loads that defines a symbol, and if its tries to load all possible
modules that provide a given symbol and fails, then returns without loading
the symbol. NativeCompiler doesn't return a positive indication that it
has a symbol until it has actually loaded the symbol.

Eventually we may not need this because we'll be loading modules partially.