Allow all Function objects to implicitly depend on a '__module_hash__' member of their globals.
In some downstream use cases, we end up with Function objects that have been created with
globals dicts that are not backed by actual modules. In this case, if those functions
reference global variables that are mutable, we have no way of differentiating two
functions with identical behavior but different globals.  This is because normally
we don't need to differentiate between them because they get access to their globals
by virtue of their reference to the global module dict, and only one such object
can exist with a given name in memory at once.  But if a function's globals are
not named, we need a way of differentiating the dicts.

To support this, we introduce a __module_hash__ variable - all TP function objects
(Entrypoint, NotCompiled, Function, and Class methods) have an implicit dependency on
this global variable.  This allows downstream code to set this value and prevent
the two functions from being the same.