Failing on
test_serialize_self_referencing_class_basic

The problem is that when we have a set of recursive types that see themselves through function
gobals, the function globals cannot be resolved to the concrete types until after identity hashing
has happened. This is a problem because identity hashing requires that the object being hashed is
stable!

The solution is to modify the Type graph so that _everything_ is contained entirely within the forward
type graph, with references to all other types (and python objects/classes) being modeled explicitly
outside of the python interpreter.  Then once the object is 'final' we can decide to merge it back into
the version that's visible to the python interpreter.