Bug repro, compiled iteration over uninitialized ListOf(bool)
Strange behavior observed when iterating through a list of bools which
was created with .setSizeUnsafe(N), so full of garbage data.

Outside of the compiler, or when using the builtin sum(), the iteration
gives the expected answer, the number of True values in the garbage data.

Inside a simple Entrypointed function to count Trues, we get a
ludicrously large answer (bigger than N).

Running the same function with some extra added statements for no
purpose (could be almost anything more than "pass"), we instead get 0.
The compiler seems to know to replace uninitialized data with False.