Expose source of original forward definitions and ensure that Classes expose members.
The problem we're having is that 'Function' and 'Entrypoint' force resolution of their
Function objects immediately.

In this case, the forward function refers to a Global whose module dict doesn't have a name
and is not yet defined. As a result, we end up forcing the resolution of the global immediately
which forces it to 'Unbound'.

This is kind of a weird test because we're constructing a global inside of a module
whose name is not known. How should we handle this? This is similar to what we do in ModuleRepresentation.

At some level, the problem is that we are conflating function types with function
instances, which happens when we mark them with @Entrypoint inside of a class definition.

When we @Function or @Entrypoint we are forcing the objects to resolve so we can instantiate them.
But this is different in a Class object where we are intending for the function to be part of
a Class. Of course, in a ModuleRepresentation, we would still want to acknowledge that the
Function is unresolved.

Ultimately the rule should be:

(1) when you @Function/@Entrypoint a function object, you get something you can use.
    * the 'future uncertainty' embedded in the object is not part of its type.
    * when you pass it into 'the compiler' we force it to resolve or be unresolved but managed by the system